use java 17 as default
Since cds-services 1.32.0 the new default is Java 17 (older
versions are still supported). Would go for 17 here as this is
a prominent sample app and we want and have to promote
the usage of Java 17 since it will be mandatory in the upcoming
cds-services 2.0 release.

Also adding Java 20 (which has just been released) to the build
pipeline configuration.