fix calculation of deduct_discount
prior to this change the new total price was calculated
with the deducted booking fee. This might be correct from
a business perspective but from a customer perspective this
would be at least surprising. Also in demos the minimal change
on the total price is so small that it's barely noticeable.